update error message show
