+++ main.ts
@@ -1,5 +1,6 @@
-import { getVQD,search } from "./search/search";
+import { search } from "./search/search";
+import { getVQD } from "./search/base";
 
-// getVQD('大语言模型').then(console.log).catch(console.error);
+getVQD('大语言模型').then(console.log).catch(console.error);
 
-search('大语言模型微调').then(console.log).catch(console.error);
+// search('大语言模型微调').then(console.log).catch(console.error);
